Finished Processor
- Fixed processor not updating the base project
- Replaced the old processor
- Added tests for inverted and fail
