+++ Test/src/main/java/com/minecrafttas/test/Test.java
@@ -1,12 +1,21 @@
 package com.minecrafttas.test;
 
 import net.fabricmc.api.ModInitializer;
+import net.minecraft.client.Minecraft;
 
 public class Test implements ModInitializer {
 
 	@Override
 	public void onInitialize() {
 		System.out.println("pancake tampered with your game");
+		//# 1.15.2
+//$$		//HELP ME
+		//# 1.14.4
+//$$		Minecraft.getInstance();
+		//# def
+//$$		Minecraft.getMinecraft();
+//$$		//Aaaaaa
+		//# end
 	}
-
+		
 }
