+++ src/test/resources/Test1/Expected4.txt
@@ -1,0 +1,11 @@
+package Test1;
+
+public class Actual {
+	//# 1.18.1
+//$$	// Code for 1.18.1 and up
+	//# 1.16.1
+//$$	// Code for 1.16.1 and up
+	//# def
+//$$	// Code for 1.14.4 and up
+	//# end
+}
