--- src/test/resources/TestFile1.java
@@ -1,7 +1,0 @@
-public class TestFile1 {
-	//# 1.18.1
-	// Code for 1.18.1 and up
-	//# 1.16.1
-	// Code for 1.16.1 and up
-	//# end
-}
