--- src/test/resources/TestFile2.java
@@ -1,14 +1,0 @@
-public class TestFile2 {
-	//# 1.18.1
-	// Code for 1.18.1 and up
-	//# 1.16.1
-	// Code for 1.16.1 and up
-	//# end
-
-	// Stuff that shouldn't be changed
-
-	//# 1.17.1
-	// Another code part
-	//# 1.16.1
-	// Weee
-}
