--- src/test/resources/TestFile3.java
@@ -1,4 +1,0 @@
-public class TestFile3 {
-	//	Minecraft mc = Minecraft.getMinecraft(); // @GetMinecraft
-	//	World level = mc.world; // @GetLevel
-}
