+++ src/test/resources/TestVersionInverted/Expected2.txt
@@ -1,0 +1,11 @@
+package TestVersionInverted;
+
+public class Actual {
+	//# 1.8.9
+//$$	// Code for 1.8.9 and below
+	//# 1.11.2
+//$$	// Code for 1.11.2 and below
+	//# def
+	// Code for 1.12.2 and below
+	//# end
+}
